Refactor Coverband::Utils::Line to Coverband::Utils::SourceFile::Line
Moved `Coverband::Utils::Line` to `lib/coverband/utils/source_file/line.rb` and renamed it to `Coverband::Utils::SourceFile::Line` to improve code organization.
Updated `lib/coverband/utils/source_file.rb` to use the new class.
Updated tests in `test/coverband/utils/source_file_test.rb` and moved `test/coverband/utils/line_test.rb` to `test/coverband/utils/source_file/line_test.rb`.
