+++ test/coverband/utils/source_file/line_test.rb
@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
-require File.expand_path("../../test_helper", File.dirname(__FILE__))
+require File.expand_path("../../../test_helper", File.dirname(__FILE__))
 
 ####
 # Thanks for all the help SimpleCov https://github.com/colszowka/simplecov-html
 # initial version of test pulled into Coverband from Simplecov 12/19/2018
 ####
-describe Coverband::Utils::Line do
+describe Coverband::Utils::SourceFile::Line do
   describe "a source line" do
     subject do
-      Coverband::Utils::Line.new("# the ruby source", 5, 3)
+      Coverband::Utils::SourceFile::Line.new("# the ruby source", 5, 3)
     end
 
     it 'returns "# the ruby source" as src' do
@@ -57,7 +57,7 @@ describe Coverband::Utils::Line do
 
   describe "A source line with coverage" do
     subject do
-      Coverband::Utils::Line.new("# the ruby source", 5, 3)
+      Coverband::Utils::SourceFile::Line.new("# the ruby source", 5, 3)
     end
 
     it "has coverage of 3" do
@@ -87,7 +87,7 @@ describe Coverband::Utils::Line do
 
   describe "A source line without coverage" do
     subject do
-      Coverband::Utils::Line.new("# the ruby source", 5, 0)
+      Coverband::Utils::SourceFile::Line.new("# the ruby source", 5, 0)
     end
 
     it "has coverage of 0" do
@@ -117,7 +117,7 @@ describe Coverband::Utils::Line do
 
   describe "A source line with no code" do
     subject do
-      Coverband::Utils::Line.new("# the ruby source", 5, nil)
+      Coverband::Utils::SourceFile::Line.new("# the ruby source", 5, nil)
     end
 
     it "has nil coverage" do
@@ -147,19 +147,19 @@ describe Coverband::Utils::Line do
 
   it "raises ArgumentError when initialized with invalid src" do
     assert_raises ArgumentError do
-      Coverband::Utils::Line.new(:symbol, 5, 3)
+      Coverband::Utils::SourceFile::Line.new(:symbol, 5, 3)
     end
   end
 
   it "raises ArgumentError when initialized with invalid line_number" do
     assert_raises ArgumentError do
-      Coverband::Utils::Line.new("some source", "five", 3)
+      Coverband::Utils::SourceFile::Line.new("some source", "five", 3)
     end
   end
 
   it "raises ArgumentError when initialized with invalid coverage" do
     assert_raises ArgumentError do
-      Coverband::Utils::Line.new("some source", 5, "three")
+      Coverband::Utils::SourceFile::Line.new("some source", 5, "three")
     end
   end
 end
